+++ python/src/mesh.cpp
@@ -197,9 +197,8 @@ void mesh(py::module& m)
       .def("dim", &dolfin::MeshEntity::dim, "Topological dimension")
       .def("mesh", &dolfin::MeshEntity::mesh, "Associated mesh")
       .def("index",
-           (std::size_t(dolfin::MeshEntity::*)() const)
-               & dolfin::MeshEntity::index,
-           "Index")
+           (std::uint32_t (dolfin::MeshEntity::*)() const)
+               & dolfin::MeshEntity::index, "Index")
       .def("global_index", &dolfin::MeshEntity::global_index, "Global index")
       .def("num_entities", &dolfin::MeshEntity::num_entities,
            "Number of incident entities of given dimension")
